Check extension codition first in case of EveryNodeReady check
There can be a scenario when infra resource is depleted while creating the first node(or min no of nodes for a worker pool). In this case, the error that comes now is `Not enough worker nodes registered` which is not correct
rather error should say infra is at limit which comes from worker extension status.

Signed-off-by: Niclas Schad <niclas.schad@stackit.de>