V1.117.0 cherry picks (#144)
* Enable AlwaysUpdate for resource manager targeting runtime cluster

* adapt gardener-node-agent to images built using ko

The changes are similar to those made by pkg/provider-local/webhook/nodeagentosc/mutator.go

* add custom configuration for `Garden`

* set insecure_skip_verify to true for cadvisor and kubelet job

Since Prometheus Targets are all down due to gardener#9716 in Gardener v1.96, we need to set insecure_skip_verify to true on cadvisor & kubelet scrape config.

Since this is a hotfix, it can be dropped whenever we have a long term
solution.

* Remove Extension Validation for Provider in ETCD-Backup

With Gardener v1.114+ the operator extensions are a required resource for the upgrade.
In our use case, we need to deactivate this validation since this is a breaking change.
We need to do a backup-s3 migration.

Drop this commit, after the migration of the extension was performed.

* Deploy proxy-protocol envoy filters if proxy-protocol is enabled for seed. (#11919)

Co-authored-by: Axel Siebenborn <axel.siebenborn@sap.com>

* `gardener-controller-manager` patches objects only when legacy seed labels have been removed (#11928)

Co-authored-by: oliver-goetz <o.goetz@sap.com>

---------

Co-authored-by: Marcel Boehm <marcel.boehm@inovex.de>
Co-authored-by: Michael Eischer <michael.eischer@inovex.de>
Co-authored-by: Tim Ebert <timebertt@gmail.com>
Co-authored-by: Duciwuci <duciwuci@gmail.com>
Co-authored-by: Gardener Prow Robot <gardener.ci.robot@gmail.com>
Co-authored-by: Axel Siebenborn <axel.siebenborn@sap.com>
Co-authored-by: oliver-goetz <o.goetz@sap.com>