[release-v1.118] Ensure extension admission webhooks validated `WorkloadIdentity`s and `Secret`s when used (#12075)
backport of https://github.com/gardener/gardener/pull/12061
Can be dropped when upgrading to g/g@v1.119

* Implement dummy `Secret`/`WorkloadIdentity` validator for provider-local

for local testing

Co-Authored-By: Tim Usner <tim.usner@sap.com>

* New webhook for syncing provider secrets

For `WorkoadIdentity`s, there is already
https://github.com/gardener/gardener/blob/b9622262d10177c185e26b4a6551414cc236ac51/plugin/pkg/global/extensionlabels/admission.go#L157-L165
which takes care of syncing the provider label based on the type in the
`spec`.

* Dry-run provider secret creation to ensure it is sane

This way, we ensure that all `*Binding` resources can be considered
"safe to be used" UNDER THE CONDITION that the provider type of the
`*Binding` matches with the provider type of the target resource (e.g.,
a `Shoot` with provider type `foo` must ensure that it only references
a `*Binding` with provider type `foo`).

This is already ensured by https://github.com/gardener/gardener/blob/2a9c566f89a6a16e654193945250f60755c560b9/plugin/pkg/shoot/validator/admission.go#L1073-L1091.
As of today, the `Shoot` API is the only one using `*Binding` resources.

Note that we don't need to do this dry-run-create for
`WorkloadIdentity`s because we already ensure that the provider types of
it and the `CredentialsBinding` match (see previous commit). This way,
we ensure that a `WorkloadIdentity` must have already successfully
passed the provider extension check when referenced in
a `CredentialsBinding` with the same type.

Co-Authored-By: Tim Usner <tim.usner@sap.com>

* New admission plugin preventing undesired finalizer removal of `*Binding`s

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Ensure provider types match

* Support `generateName` field

When `CredentialsBinding`s or `SecretBinding`s are created with `generateName`, the `metadata.name` field is not populated, causing the admission plugin to miss validating that the providers between the binding and the shoot match.

In addition, the `ResourceReferenceManager` needs to be a validating admission plugin in order to see the changes applied by `PrepareForCreate`.

* Handle unset `provider` field in `SecretBinding`s

* Address PR review feedback

---------

Co-authored-by: rfranzke <rafael.franzke@sap.com>
Co-authored-by: Tim Usner <tim.usner@sap.com>